Better bookmark install flow, minor improvements
Put the `:javascript:` part of the bookmark in the file,
it doesn't cause issues in IDE. Newlines in bookmarks are not a problem.

Attempt to locate the last edit time of a message on GitHub.
It doesn't necessarily lead to a better result,
but should overall result in more matches.

Clear output before calling GH API